break out telemetry and playbook
